dev-requirements.txt
pytest-ckan
pytest_factoryboy
pytest-cov